requirements/base.txt
Django==4.1.7
django-environ==0.10.0
djangorestframework==3.14.0
drf-yasg==1.21.5
pyyaml==6.0
requests==2.28.2
